+++ lib/src/file_context.dart
@@ -79,8 +79,7 @@ class FileContext {
     var buffer = StringBuffer();
     for (var error in result.errors) {
       var location = result.lineInfo.getLocation(error.offset);
-      // ignore: deprecated_member_use
-      buffer.writeln('  ${error.errorCode.name}: ${error.message} - '
+      buffer.writeln('  ${error.diagnosticCode.lowerCaseName}: ${error.message} - '
           '${location.lineNumber}:${location.columnNumber}');
     }
     throw ArgumentError(