+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- **Breaking:** Require analyzer `^14.0.0` and Dart SDK `^3.11.0`.
+
 ## [1.3.0](https://github.com/Workiva/dart_codemod/compare/1.2.0...1.3.0)
 
 - Allow up to analyzer 10